+++ pom.xml
@@ -8,7 +8,7 @@
 		<version>2.4.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.lokesh.samples</groupId>
+	<groupId>com.lokesh.java.samples</groupId>
 	<artifactId>CurrencyRatesPublisher</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>CurrencyRatesPublisher</name>
