+++ src/main/java/com/crp/components/CurrencyRatesPublisherApplication.java
@@ -1,4 +1,4 @@
-package com.lokesh.publisher;
+package com.crp.components;
 
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
@@ -14,8 +14,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CurrencyPublisherApplication implements CommandLineRunner {
-	private static Logger Log = Logger.getLogger(CurrencyPublisherApplication.class.getName());
+public class CurrencyRatesPublisherApplication implements CommandLineRunner {
+	private static Logger Log = Logger.getLogger(CurrencyRatesPublisherApplication.class.getName());
 	private static ApplicationContext appCxt = null;
 	
 	@Autowired
@@ -23,7 +23,7 @@ public class CurrencyPublisherApplication implements CommandLineRunner {
 	
 	public static void main(String[] args) {
 		Log.info("Starting CurrencyPublisherApplication");
-		SpringApplication app = new SpringApplication(CurrencyPublisherApplication.class);
+		SpringApplication app = new SpringApplication(CurrencyRatesPublisherApplication.class);
 		app.setBannerMode(Mode.OFF);
 		Log.info("Switched off application banner");
 		appCxt = app.run(args);
@@ -33,8 +33,8 @@ public class CurrencyPublisherApplication implements CommandLineRunner {
 	@Override
 	public void run(String... args) throws Exception {
 		Log.info("Running CommandLineRunner run method");
-		CurrencyPublisherThreadsPool publisherThreads 
-		      = new CurrencyPublisherThreadsPool(env.getProperty("currency.sdate"));
+		CurrencyRatesPublisherThreadsPool publisherThreads 
+		      = new CurrencyRatesPublisherThreadsPool(env.getProperty("currency.sdate"));
 		
 		publisherThreads.executeCurrencyPulisherThreads();
 		Log.info("Return from CommandLineRunner run method");
